python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cornserve"
description = "Easy, fast, and scalable agentic AI"
readme = "README.md"
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
]
requires-python = ">=3.12"
dependencies = [
    "grpcio-tools",
    "tyro",
    "kubernetes_asyncio",
]
dynamic = ["version"]

[project.optional-dependencies]
dev = ["grpcio-tools", "black", "pyright", "ruff"]

[tool.setuptools.dynamic]
version = { attr = "cornserve.__version__" }

[tool.ruff]
line-length = 120
pydocstyle.convention = "google"
select = [
  "E",   # pycodestyle error
  "F",   # pyflakes
  "D",   # pydocstyle
  "PL",  # pylint
  "N",   # pep8-naming
  "B",   # flake8-bugbear (detects likely bugs)
  "G",   # flake8-logging-format (complains about logging)
  "SIM", # flake8-simplify (suggests code simplifications)
]

[tool.pyright]
exclude = [
    "**/*_pb2.py",
    "**/*_pb2_grpc.py",
]
